split backend OpenAI adapter folders
