+++ apps/backend/src/chat/openai/loop/index.ts
@@ -1,0 +1,12 @@
+export {
+  buildPromptCacheKey,
+  CHAT_RUN_MAX_TOOL_CALL_MODEL_CALLS,
+  startOpenAILoop,
+  startOpenAILoopWithDeps,
+  startOpenAILoopWithDeps as startOpenAILoopWithDependencies,
+} from "./loop";
+export type {
+  OpenAILoopCompletion,
+  OpenAILoopEventSink,
+  StartOpenAILoopParams,
+} from "./loop";
